requirements.txt
django~=5.0
djangorestframework~=3.14.0
psycopg2~=2.9.9
drf-spectacular>=0.25.1,<0.27.0






